Reduce median calldata size on timestamp

+++ src/mosm.sol
@@ -96,7 +96,7 @@ contract Mosm is LibNote {
         return (val, val > 0);
     }
 
-    function recover(uint256 val_, uint256 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
+    function recover(uint256 val_, uint32 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
         return ecrecover(
             keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(abi.encodePacked(val_, age_, wat)))),
             v, r, s
@@ -104,14 +104,14 @@ contract Mosm is LibNote {
     }
 
     function median_poke(
-        uint256[] calldata val_, uint256[] calldata age_,
+        uint256[] calldata val_, uint32[] calldata age_,
         uint8[] calldata v, bytes32[] calldata r, bytes32[] calldata s) external
     {
         require(val_.length == bar, "Median/bar-too-low");
 
         uint256 bloom = 0;
         uint256 last = 0;
-        uint256 snooze = age;
+        uint32 snooze = age;
 
         for (uint i = 0; i < val_.length; i++) {
             // Validate the values were signed by an authorized oracle
@@ -306,7 +306,7 @@ contract Mosm is LibNote {
 // contract MosmETHUSD is Mosm {
 //     bytes32 public constant wat = "ETHUSD";
 //
-//     function recover(uint256 val_, uint256 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
+//     function recover(uint256 val_, uint32 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
 //         return ecrecover(
 //             keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(abi.encodePacked(val_, age_, wat)))),
 //             v, r, s
